+++ docs/docs/resources/changelog.md
@@ -33,6 +33,8 @@ During the alpha period, things might break! We take breaking changes very serio
 - Added the [`.wa-cloak` utility](/docs/utilities/fouce) to prevent FOUCE
 - Added the [`allDefined()` utility](/docs/usage/#all-defined) for awaiting component registration
 - Simplified `<wa-breadcrumb-item>` by removing the `base` CSS part
+- Simplified `<wa-menu-item>` and `<wa-menu-label>` by removing the `base` CSS part
+- Added slots to `checked-icon` and `submenu-icon` in `<wa-menu-item>` so custom icons can be used
 
 ### Bug fixes